server/requirements.txt
--no-binary psycopg2

Django==2.0.2 --hash=sha256:7c8ff92285406fb349e765e9ade685eec7271d6f5c3f918e495a74768b765c99
pytz==2018.3 --hash=sha256:07edfc3d4d2705a20a6e99d97f0c4b61c800b8232dc1c04d87e8554f130148dd
psycopg2==2.7.4 --hash=sha256:8bf51191d60f6987482ef0cfe8511bbf4877a5aa7f313d7b488b53189cf26209
djangorestframework==3.7.7 --hash=sha256:1f6baf40ed456ed2af6bd1a4ff8bbc3503cebea16509993aea2b7085bc097766
django-cors-headers==2.1.0 --hash=sha256:4e02be61ffaaab5917f1fd7cc3c305c4fb7ccd0156a649c96f49bc0a09c5f572
djoser==1.1.5 --hash=sha256:5348b153d887dfea86e0ecafce1c63588705e8e74d907ba5b1d54ab94bf85ab1
django-templated-mail==1.1.1 --hash=sha256:f7127e1e31d7cad4e6c4b4801d25814d4b8782627ead76f4a75b3b7650687556  # via djoser

# Import/export
django-import-export==0.7.0 --hash=sha256:7e7ebeb40702eafeb2e770914c01b9961063f472b3b395eeffbea5f39efa7257
django-filter==1.1.0 --hash=sha256:ea204242ea83790e1512c9d0d8255002a652a6f4986e93cee664f28955ba0c22
diff-match-patch==20121119 --hash=sha256:9dba5611fbf27893347349fd51cc1911cb403682a7163373adacc565d11e2e4c  # via django-import-export
tablib==0.12.1 --hash=sha256:b8cf50a61d66655229993f2ee29220553fb2c80403479f8e6de77c0c24649d87  # via django-import-export
odfpy==1.3.6 --hash=sha256:6bcaf3b23aa9e49ed8c8c177266539b211add4e02402748a994451482a10cb1b  # via tablib
openpyxl==2.4.9 --hash=sha256:95e007f4d121f4fd73f39a6d74a883c75e9fa9d96de91d43c1641c103c3a9b18  # via tablib
et_xmlfile==1.0.1 --hash=sha256:614d9722d572f6246302c4491846d2c393c199cfa4edc9af593437691683335b  # via openpyxl
jdcal==1.3 --hash=sha256:b760160f8dc8cc51d17875c6b663fafe64be699e10ce34b6a95184b5aa0fdc9e  # via openpyxl
unicodecsv==0.14.1 --hash=sha256:018c08037d48649a0412063ff4eda26eaa81eff1546dbffa51fa5293276ff7fc  # via tablib
xlrd==1.1.0 --hash=sha256:83a1d2f1091078fb3f65876753b5302c5cfb6a41de64b9587b74cefa75157148  # via tablib
xlwt==1.3.0 --hash=sha256:a082260524678ba48a297d922cc385f58278b8aa68741596a87de01a9c628b2e  # via tablib
PyYAML==3.12 --hash=sha256:592766c6303207a20efc445587778322d7f73b161bd994f227adaa341ba212ab  # via tablib
